requirements.txt
flask
flasgger
pymongo # If you are using MongoDB
python-dotenv # If you are using environment variables
